Don't prune non-existent snapshots for non-recursive streams #2907 [dweeezil]
A send stream with properties contains an nvlist "fss" describing
the source snapshots and their properties. As of commit
0574855, this list may not include
all the snapshots on the source. In particular, it only includes the
snapshots which actually are sent and, furthermore, it is only sent when
either "zfs send -p" or "zfs send -R" are used.

The snapshot pruning code enabled under "zfs receive -F" depends on a
complete fss[id[snaps]] list. If it's not complete, the filesystem can
be damaged at worst or, at best, left in it's renamed state (renamed to
"-recv--".

This commit suppresses the snapshot pruning for non-recursive streams
since they still include the complete list of snapshots in fss[id[snaps]].

https://github.com/zfsonlinux/zfs/pull/2907