Restructure per-filesystem reclaim
Originally when the ARC prune callback was introduced the idea was
to register a single callback for the ZPL.  The ARC could invoke this
call back if it needed the ZPL to drop dentries, inodes, or other
cache objects which might be pinning buffers in the ARC.  The ZPL
would iterate over all ZFS super block and perform the reclaim.

For the most part this design have worked well but due to limitations
in 2.6.35 and earlier kernels there were some problems.  This patch
is desgined to address those issues.

1) iterate_supers_type() is not provided by all kernels which makes
it impossible to safely iterate over all zpl_fs_type filesystems in
a single callback.  The most straight forward and portable way to
resolve this is to register a callback per-filesystem during mount.
The arc_*_prune_callback() functions have always supported multiple
callbacks so this is functionally a very small change.

2) Commit 050d22b removed the non-portable shrink_dcache_memory()
and shrink_icache_memory() functions and didn't replace them with
equivilant functionality.  This meant that for Linux 3.1 and older
kernels the ARC had no mechanism to drop dentries and inodes from
the caches if needed.  This patch adds the missing functionalilty
by calling shrink_dcache_parent() and invalidate_inodes() for these
kernels.  This will result in all unused cache entries being dropped
which is a bit heavy handed but it's the only interface available.

3) A zpl_drop_inode() callback is registered for kernels older than
2.6.35 which do not support .evict_inode.  This ensures that when
the last reference on an inode is dropped it is immediately removed
from the cache.  If this isn't done than inode can end up on the
global unused LRU with no mechanism available to ZFS to drop them.
Since the ARC buffers are not dropped the hottest inodes can still
be recreated without performing disk IO.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Isssue #3160