Restore dbuf_write_ready() early return #3158
Commit b0bc7a8 reworked the dbuf_write_ready() function to remove the
early return when a hole is detected. This does simplify the code but
it means additional contention of the db->db_mtx lock which can impact
performance.

Specificially the tiotest benchmark shows a significant performance
regression because it reads/writes zero-filled buffers which are converted
to holes. Without the early return optimization the additional lock
contention on the dbufs is significant. Cached read performance for
holes is cut in half. The performance of non-hole block pointers
remains roughly the same as expected.

Signed-off-by: Brian Behlendorf behlendorf1@llnl.gov
Issue #3106