Linux 3.18 compat: Snapshot automounting
Commit torvalds/linux@bafc9b7 caused snapshots automounted by ZFS to
be immediately unmounted when the dentry was revalidated.  This was
a consequence of ZFS invaliding all snapdir dentries to ensure that
negative dentries didn't mask new snapshots.  This patch modifies the
behavior such that only negative dentries are invalidated.  This solves
the issue and may result in a performance improvement.

In addition, the MNT_SHRINKABLE flag is now set after the automount
succeeds by travesing down in to the snapshot.  This is much cleaner
than being forced to do it post in a getattr.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Issue #3030