Fix misuse of input argument in traverse_visitbp
In traverse_visitbp(), the input argument dnp is modified in the middle to
point to a temporary buffer. Originally this doesn't matter, because nobody
use it after that. However, in fbeddd6 a piece of code is added referencing
dnp after the modification, creating a possible bug.

We fix this by creating a new local variable cdnp so we don't touch the input
argument.

Signed-off-by: Chunwei Chen <tuxoko@gmail.com>